requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
docopt==0.6.2
idna==3.4
numpy==1.21.6
oauthlib==3.2.1
pandas==1.3.5
py4j==0.10.9.5
pyspark==3.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
tweepy==4.10.1
urllib3==1.26.12
yarg==0.1.9
